upgrade paas2 to support both python2/3
